Improved integraton test and some cleanup. Lets use as little hoodiecrow api as possible, so we can replace it easily someday.
