+++ package.json
@@ -33,7 +33,6 @@
     "url": "https://github.com/chadxz/imap-simple.git"
   },
   "dependencies": {
-    "hoodiecrow-imap": "^2.1.0",
     "iconv-lite": "~0.4.13",
     "imap": "^0.8.18",
     "nodeify": "^1.0.0",
@@ -42,6 +41,7 @@
     "uuencode": "0.0.4"
   },
   "devDependencies": {
+    "hoodiecrow-imap": "^2.1.0",
     "chai": "^3.5.0",
     "istanbul": "^0.4.5",
     "jscs": "^3.0.7",
